autoresponse/pyproject.toml
[project]
name = "autoresponse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.2.0",
    "google-adk>=1.28.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
